+++ Runtime/SdkVersion.cs
@@ -7,6 +7,6 @@ namespace Unity.Services.Analytics
         /// This value should always match what is in the package.json file for this package,
         /// to provide accurate SDK versions to the analytics data classes.
         /// </summary>
-        public static readonly string SDK_VERSION = "5.1.1";
+        public static readonly string SDK_VERSION = "6.0.1";
     }
 }